Move `inventory` subcommands to `cli/inventory.py`
